Do not redirect on POST
